Rename validator.py
